backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
bcrypt
email-validator
httpx
pyjwt
pytest
pytest-asyncio
